move providers to hooks dir
