+++ src/main.tsx
@@ -1,31 +1,31 @@
-import React from "react";
-import { createRoot } from "react-dom/client";
-import { configureChains, createConfig, WagmiConfig } from "wagmi";
+import React from 'react';
+import { createRoot } from 'react-dom/client';
+import { configureChains, createConfig, WagmiConfig } from 'wagmi';
 
-import { hardhat, polygon } from "wagmi/chains";
-import "./shims";
+import { hardhat, polygon } from 'wagmi/chains';
+import './shims';
 
-import { HashRouter } from "react-router-dom";
+import { HashRouter } from 'react-router-dom';
 
-import { CoinbaseWalletConnector } from "wagmi/connectors/coinbaseWallet";
-import { InjectedConnector } from "wagmi/connectors/injected";
-import { MetaMaskConnector } from "wagmi/connectors/metaMask";
-import { alchemyProvider } from "wagmi/providers/alchemy";
-import { publicProvider } from "wagmi/providers/public";
-import App from "./components/App";
-import { ModalProvider } from "./context/ModalContext";
-import { ToastProvider } from "./context/ToastContext";
-import { TransactionDisplayProvider } from "./context/TransactionDisplayContext";
-import "./css/index.css";
+import { CoinbaseWalletConnector } from 'wagmi/connectors/coinbaseWallet';
+import { InjectedConnector } from 'wagmi/connectors/injected';
+import { MetaMaskConnector } from 'wagmi/connectors/metaMask';
+import { alchemyProvider } from 'wagmi/providers/alchemy';
+import { publicProvider } from 'wagmi/providers/public';
+import App from './components/App';
+import './css/index.css';
+import { ModalProvider } from './hooks/ModalContext';
+import { ToastProvider } from './hooks/ToastContext';
+import { TransactionDisplayProvider } from './hooks/TransactionDisplayContext';
 
-export type IViteMode = "production" | "development" | "testing" | undefined;
+export type IViteMode = 'production' | 'development' | 'testing' | undefined;
 
 const env = import.meta.env.MODE as IViteMode;
 
-if (env === undefined) throw new Error("NODE_ENV not set!");
+if (env === undefined) throw new Error('NODE_ENV not set!');
 
-const container = document.getElementById("root");
-if (!container) throw new Error("no root element found!");
+const container = document.getElementById('root');
+if (!container) throw new Error('no root element found!');
 
 const apiKey = import.meta.env.VITE_ALCHEMY_ID as string;
 
@@ -41,14 +41,14 @@ const config = createConfig({
     new CoinbaseWalletConnector({
       chains,
       options: {
-        appName: "wagmi",
+        appName: 'wagmi',
       },
     }),
 
     new InjectedConnector({
       chains,
       options: {
-        name: "Injected",
+        name: 'Injected',
         shimDisconnect: true,
       },
     }),
